LayoutTests/platform/ios-sim-deprecated/fast/loader/subframe-removes-itself-expected.txt

frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFailLoadWithError
main frame - didFinishLoadForFrame
This tests a subframe that removes itself to make sure a delegate message to indicate the interrupted load fires.
